requirements.txt
# IRTS Python Harvest System Dependencies

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# HTTP requests
requests>=2.31.0

# XML/JSON parsing
lxml>=4.9.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Type hints
typing-extensions>=4.8.0
